64-bit changes: ioctl cmd -> u_long, some protos.  some style, return vals.

+++ sys/sys/resource.h
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)resource.h	8.3 (Berkeley) 07/10/94
+ *	@(#)resource.h	8.4 (Berkeley) 01/09/95
  */
 
 #ifndef _SYS_RESOURCE_H_
@@ -78,7 +78,7 @@ struct rlimit {
 /* Load average structure. */
 struct loadavg {
 	fixpt_t	ldavg[3];
-	int32_t fscale;
+	long	fscale;
 };
 
 #ifdef KERNEL
